Refactor transportation booking components and update routes for improved structure and clarity
