+++ lib/features/my_bookings/view/residence_booking/screens/details_screen.dart
